refactor: move health check to HealthCheckManager and HealthCheckRouter
