+++ server/src/routes/middleware/RouteMap.js
@@ -7,6 +7,7 @@ const { RES_LOCALS } = require("./constant");
 const AuthModel = require("../../models/AuthModel");
 const GithubAnalyticsRouter = require("../router/gitHubAnalyticsRouter");
 const PlagiarismRouter = require("../router/plagiarismRouter");
+const HealthCheckRouter = require("../router/healthCheckRouter");
 const Db = require("../../models/libs/Db");
 const redis = require("../../config/redis");
 const { JWT_SECRET_KEY } = require("../../../constants");
@@ -24,31 +25,7 @@ class RouteMap {
     // 🔓 OPEN ROUTES (NO JWT)
     // app.use("/open/api/auth",Authrouter);
 
-    openrouter.get("/health", async (req, res) => {
-      const health = {
-        server: "up",
-        database: "down",
-        redis: "down",
-        timestamp: new Date().toISOString()
-      };
-
-      try {
-        await Db.getQueryBuilder().raw('SELECT 1');
-        health.database = "up";
-      } catch (err) {
-        console.error("Health check database error:", err);
-      }
-
-      try {
-        await redis.ping();
-        health.redis = "up";
-      } catch (err) {
-        console.error("Health check redis error:", err);
-      }
-
-      const isHealthy = health.database === "up" && health.redis === "up";
-      res.status(isHealthy ? 200 : 503).json(health);
-    });
+    openrouter.use("/health", HealthCheckRouter);
       
 
     openrouter.use("/github", GithubAnalyticsRouter);
